requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
more-itertools==6.0.0
pluggy==0.8.1
py==1.8.0
pytest==4.3.0
PyYAML==4.2b4
six==1.12.0
